Fix bug. Deleted resources can reference deleted ones.
Say you have a stack with three resources A, B, C, and B `Ref`s C. If
you remove both B and C from the stack, moto will break -- it tries to
resolve the reference to C using the _end state of the stack_ which
includes only A (and does not include C).

Fix this by not updating the resource map of self until after the
removed resources have been deleted.