Fix bug. Reset self._template during stack update.
update calls diff. diff mutates self by calling load_parameters. but the
latter reads the value of self._template, which is out of date when we
are doing an update.

set self._template to the given template in diff.

also, change the following identifiers to be more clear (since they
represent not templates but resource maps):

* `old_template` -> `old_template_resources`
* `new_template` -> `new_template_resources`